Various improvements coming out of stress testing
