+++ client-interface/src/lib.rs
@@ -98,6 +98,9 @@ impl SubstrateNetwork {
         }
     }
 
+    // TODO The substrate node endpoints are proving to be very unreliable when it comes to looking
+    //  up old blocks. This is happening to both Rococo and Kusama. Replace uses of this function
+    //  with the equivalent subscan one.
     pub async fn get_extrinsic(
         &self,
         location: ExtrinsicLocation
@@ -392,7 +395,6 @@ pub fn account_to_subxt_multi_address(account: AccountId32) -> SubxtMultiAddress
 pub struct ServiceInfo {
     pub proxy_account: AccountId32,
     pub network_name: String,
-    pub node_endpoint: String,
     #[serde(with = "common::serde_over_hex_scale")]
     pub attestation_bundle: AttestationBundle,
     pub version: String
@@ -439,7 +441,6 @@ mod tests {
         let service_info = ServiceInfo {
             proxy_account: dev::alice().public_key().0.into(),
             network_name: "polkadot".to_string(),
-            node_endpoint: "wss://polkadot.api.onfinality.io/public-ws".to_string(),
             attestation_bundle: AttestationBundle {
                 attested_data: AttestedData {
                     genesis_hash: random::<[u8; 32]>().into(),
@@ -459,7 +460,6 @@ mod tests {
             json!({
                 "proxy_account": "5GrwvaEF5zXb26Fz9rcQpDWS57CtERHpNehXCPcNoHGKutQY",
                 "network_name": "polkadot",
-                "node_endpoint": "wss://polkadot.api.onfinality.io/public-ws",
                 "attestation_bundle": hex::encode(&service_info.attestation_bundle.encode()),
                 "version": "1.0.0"
             })
