+++ service/Cargo.toml
@@ -28,6 +28,7 @@ serde.workspace = true
 reqwest.workspace = true
 aws-config.workspace = true
 aws-sdk-dynamodb.workspace = true
+aws-smithy-runtime-api.workspace = true
 
 [target.'cfg(target_os = "linux")'.dependencies]
 tokio-vsock.workspace = true
